Move the `metadata.options` namespace from `Process` to `CalcJob`
The `metadata.options` namespace only applies to the `CalcJob` process
class, yet it is defined on the `Process` base class. This means that,
process builders for `WorkChains`, for example, will also display this as
an available namespace, even though it shouldn't. It may seem harmless,
but having an apparently namespace available that does not actually do
anything can be confusing.
