+++ aiida/engine/processes/__init__.py
@@ -17,8 +17,9 @@ from .builder import *
 from .calcjobs import *
 from .exit_code import *
 from .functions import *
+from .ports import *
 from .process import *
 from .workchains import *
 
-__all__ = (
-    builder.__all__ + calcjobs.__all__ + exit_code.__all__ + functions.__all__ + process.__all__ + workchains.__all__)
+__all__ = (builder.__all__ + calcjobs.__all__ + exit_code.__all__ + functions.__all__ + ports.__all__ + process.__all__
+           + workchains.__all__)
